Basic command line functionality is working.
This program has the bare minimum functionality.
If you pass the twitch tracker URL, then it starts to download the file
to a file called `index.ts`.
It's still relatively slow at downloading compared to `yt-dlp`.
Also, you can't continue a download if it is interrupted.
The text output is also somewhat ugly.