Added streamscharts.com and added writing to disk
streamscharts.com doesn't provide the exact start time of a stream.
It only provides the minute.
So I need to check every second of that minute it provides.
The main benefit of streamscharts.com is that it provides video ids
for streams that have restarted due to a disconnection.
twitchtracker.com doesn't seem to provide these video ids.

You can also write to disk with the `--write` flag.
This writes the file to the Downloads folder.